+++ src/main/java/com/example/codionbe/global/common/exception/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
 package com.example.codionbe.global.common.exception;
 
-import com.example.codionbe.global.common.CustomException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
